refactor: move changed file picker data def
